clean up, fix warnings
