+++ src/Microsoft.DotNet.Interactive.Http/ClientExecutionRouter.cs
@@ -7,9 +7,7 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
-using Microsoft.DotNet.Interactive.Documents;
-using Microsoft.DotNet.Interactive.Extensions;
-using Microsoft.DotNet.Interactive.Formatting;
+using Microsoft.DotNet.Interactive.Http.Utility;
 
 namespace Microsoft.DotNet.Interactive.Http
 {
